move vfolder related types to data.types module
